+++ pyoffers/models/conversion.py
@@ -24,5 +24,5 @@ class ConversionManager(ModelManager):
     def find_by_id(self, id):
         return self._call('findById', id=id)
 
-    def find_all(self, filters=None, sort=None, limit=None, page=None):
-        return self._call('findAll', filters=filters, sort=sort, limit=limit, page=page, single_result=False)
+    def find_all(self, sort=None, limit=None, page=None, **kwargs):
+        return self._call('findAll', filters=kwargs, sort=sort, limit=limit, page=page, single_result=False)