+++ pyoffers/utils.py
@@ -13,6 +13,20 @@ def prepare_query_params(**kwargs):
         if sub_value is not None
     ]
 
+OPERATORS = {
+    'ne': 'NOT_EQUAL_TO',
+    'lt': 'LESS_THAN',
+    'lte': 'LESS_THAN_OR_EQUAL_TO',
+    'gt': 'GREATER_THAN',
+    'gte': 'GREATER_THAN_OR_EQUAL_TO',
+    'like': 'LIKE',
+    'not_like': 'NOT_LIKE',
+    'null': 'NULL',
+    'not_null': 'NOT_NULL',
+    'true': 'TRUE',
+    'false': 'FALSE',
+}
+
 
 @singledispatch
 def expand(value, key):
@@ -21,8 +35,16 @@ def expand(value, key):
 
 @expand.register(dict)
 def expand_dict(value, key):
-    for inner_key, inner_value in value.items():
-        yield '%s[%s]' % (key, inner_key), inner_value
+    for dict_key, dict_value in value.items():
+        if isinstance(dict_value, (list, tuple, set)):
+            for sub_value in dict_value:
+                yield '%s[%s][]' % (key, dict_key), sub_value
+        else:
+            try:
+                field_name, operator = dict_key.split('__')
+                yield '%s[%s][%s]' % (key, field_name, OPERATORS[operator]), dict_value
+            except ValueError:
+                yield '%s[%s]' % (key, dict_key), dict_value
 
 
 @expand.register(list)