+++ tests/models/test_conversion.py
@@ -27,25 +27,23 @@ def test_find_all(api):
 
 
 def test_find_all_not_paginated(api):
-    result = api.conversions.find_all(filters={'offer_id': 9})
+    result = api.conversions.find_all(offer_id=9)
     assert len(result) == 2
     assert all(isinstance(item, Conversion) for item in result)
 
 
 def test_find_all_single_result(api):
-    result = api.conversions.find_all(
-        filters={'offer_id': [7], 'source': 'FP', 'status': 'approved', 'session_ip': '127.0.0.1'}
-    )
+    result = api.conversions.find_all(offer_id=7, source='FP', status='approved', session_ip='127.0.0.1')
     assert len(result) == 1
     assert isinstance(result[0], Conversion)
 
 
 def test_find_all_not_paginated_not_found(api):
-    assert api.conversions.find_all(filters={'offer_id': 11}) == []
+    assert api.conversions.find_all(offer_id=11) == []
 
 
 def test_find_all_not_found(api):
-    assert api.conversions.find_all(filters={'goal_id': 100000}, limit=1) == []
+    assert api.conversions.find_all(goal_id=100000, limit=1) == []
 
 
 def test_update_success(conversion):